requirements.txt
fastapi[standard]
openai
pandas
python-dotenv
pymongo